Don't use PPID to get parent pid
We cannot rely on $PPID to get the parent pid of the running process
because in Docker 1.11, there is a short-lived process that is the
parent of the cotainer when it launches, but then that process dies and
the container process's parent changes but the value of $PPID does not
change.

We first sleep to ensure we give the short-lived process time to die
before looking up the parent.